fix: 同一ポートで複数サーバー設定の対応
サーバーソケット作成前にポートの重複を確認し、必要数しか生成しないように修正

+++ srcs/server/socket.hpp
@@ -8,7 +8,7 @@
 class Socket {
  public:
   int listenfd_;
-  int port_;
+  size_t port_;
   struct sockaddr_in serv_addr_;
 
  public:
